DateTypes

+++ backend/main.py
@@ -8,7 +8,7 @@ from fastapi import FastAPI, status, HTTPException
 from fastapi.middleware.cors import CORSMiddleware
 
 from model import Team, TeamMember, get_unique_countries, DayType
-from pydantic import BaseModel, Field, computed_field
+from pydantic import BaseModel, Field, computed_field, validator
 from pydantic.functional_validators import field_validator
 from fastapi.responses import RedirectResponse
 import logging
@@ -44,13 +44,22 @@ class DayTypeWriteDTO(BaseModel):
 class DayTypeReadDTO(DayTypeWriteDTO):
     id: str = Field(None, alias='_id')
 
+    @classmethod
+    def from_mongo_reference_field(cls, day_type_document_reference):
+        if day_type_document_reference:
+            day_type_document = DayType.objects.get(id=day_type_document_reference)
+            return cls(_id=str(day_type_document.id),
+                       name=day_type_document.name,
+                       color=day_type_document.color)
+        return None
+
 
 class TeamMemberWriteDTO(BaseModel):
     name: str
     country: str
     email: Optional[str] = None
     phone: Optional[str] = None
-    vac_days: List[datetime.date] = []
+    available_day_types: List[DayTypeReadDTO] = []
 
     @field_validator("country")
     @classmethod
@@ -59,21 +68,38 @@ class TeamMemberWriteDTO(BaseModel):
             return value
         raise ValueError("Invalid country name")
 
+    @field_validator('email')
+    @classmethod
+    def empty_email_to_none(cls, v):
+        return None if v == "" else v
+
 
 class TeamMemberReadDTO(TeamMemberWriteDTO):
     uid: str
+    days: Dict[str, List[DayTypeReadDTO]] = Field(default_factory=dict)
 
     @computed_field
     @property
     def vacation_days_by_year(self) -> Dict[int, int]:
         vac_days_count = defaultdict(int)
-        for day in self.vac_days:
-            vac_days_count[day.year] += 1
+        vacation_day_type = mongo_to_pydantic(DayType.objects(name="Vacation").first(), DayTypeReadDTO)
+        for date_str, day_types in self.days.items():
+            date = datetime.datetime.strptime(date_str, "%Y-%m-%d").date()
+            if vacation_day_type in day_types:
+                vac_days_count[date.year] += 1
         return dict(vac_days_count)
+
+    @field_validator('days')
+    @classmethod
+    def convert_days(cls, v):
+        if v:
+            return [DayTypeReadDTO.from_mongo_reference_field(day_type) for day_type in v]
+        return v
 
 
 class TeamWriteDTO(BaseModel):
     name: str
+    available_day_types: List[DayTypeReadDTO] = []
 
 
 class TeamReadDTO(TeamWriteDTO):
@@ -113,17 +139,7 @@ def get_holidays(year: int = datetime.datetime.now().year):
 @app.get("/")
 def read_root(year: int = datetime.datetime.now().year):
     return {"teams": list(map(lambda x: mongo_to_pydantic(x, TeamReadDTO), Team.objects.order_by("name"))),
-            "holidays": get_holidays(year)}
-
-
-@app.post("/teams/{team_id}/members/{team_member_id}/vac_days/")
-def add_vac_days(team_id: str, team_member_id: str, vac_days: List[datetime.date]):
-    team = Team.objects(id=team_id).first()
-    team_member = team.team_members.get(uid=team_member_id)
-    vac_days = set(vac_days)
-    team_member.vac_days = list(set(team_member.vac_days) | vac_days)
-    team.save()
-    return {"team": mongo_to_pydantic(team, TeamReadDTO)}
+            "holidays": get_holidays(year)} | get_all_day_types()
 
 
 @app.post("/teams/{team_id}/members/")
@@ -138,7 +154,8 @@ def add_team_member(team_id: str, team_member_dto: TeamMemberWriteDTO):
 
 @app.post("/teams/")
 def add_team(team_dto: TeamWriteDTO):
-    team = Team(name=team_dto.name).save()
+    team_data = team_dto.model_dump()
+    team = Team(**team_data).save()
     return {"team_id": str(team.id)}
 
 
@@ -158,21 +175,12 @@ def delete_team_member(team_id: str, team_member_id: str):
     return RedirectResponse(url="/", status_code=status.HTTP_303_SEE_OTHER)
 
 
-@app.delete("/teams/{team_id}/members/{team_member_id}/vac_days/")
-def delete_vac_days(team_id: str, team_member_id: str, vac_days: List[datetime.date]):
-    team = Team.objects(id=team_id).first()
-    team_member = team.team_members.get(uid=team_member_id)
-    vac_days = set(vac_days)
-    team_member.vac_days = list(set(team_member.vac_days) - vac_days)
-    team.save()
-    return {"team": mongo_to_pydantic(team, TeamReadDTO)}
-
-
 @app.put("/teams/{team_id}")
 def update_team(team_id: str, team_dto: TeamWriteDTO):
     team = Team.objects(id=team_id).first()
     if team:
         team.name = team_dto.name
+        team.available_day_types = team_dto.available_day_types
         team.save()
         return {"team": mongo_to_pydantic(team, TeamReadDTO)}
     else:
@@ -185,7 +193,7 @@ def update_team_member(team_id: str, team_member_id: str, team_member_dto: TeamM
     if not team:
         raise HTTPException(status_code=404, detail="Team not found")
 
-    team_member = team.team_members.get(uid=team_member_id)
+    team_member: TeamMember = team.team_members.get(uid=team_member_id)
     if not team_member:
         raise HTTPException(status_code=404, detail="Team member not found")
 
@@ -193,13 +201,14 @@ def update_team_member(team_id: str, team_member_id: str, team_member_dto: TeamM
     team_member.country = team_member_dto.country
     team_member.email = team_member_dto.email if team_member_dto.email else None
     team_member.phone = team_member_dto.phone
+    team_member.available_day_types = team_member_dto.available_day_types
 
     team.save()
     return {"team": mongo_to_pydantic(team, TeamReadDTO)}
 
 
-@app.put("/teams/{team_id}/members/{team_member_id}/vac_days")
-def update_vac_days(team_id: str, team_member_id: str, vac_days: List[datetime.date]):
+@app.post("/teams/{team_id}/members/{team_member_id}/days")
+def add_days(team_id: str, team_member_id: str, new_days: Dict[str, List[str]]):
     team: Team = Team.objects(id=team_id).first()
     if not team:
         raise HTTPException(status_code=404, detail="Team not found")
@@ -208,13 +217,52 @@ def update_vac_days(team_id: str, team_member_id: str, vac_days: List[datetime.d
     if not team_member:
         raise HTTPException(status_code=404, detail="Team member not found")
 
-    team_member.vac_days = vac_days
+    # Updating days field with new entries
+    for date_str, day_type_ids in new_days.items():
+        day_types = [DayType.objects(id=day_type_id).first() for day_type_id in day_type_ids]
+        if date_str in team_member.days:
+            # Add new day types to the existing list for this date
+            team_member.days[date_str].extend(day_types)
+        else:
+            # Create a new entry for this date
+            team_member.days[date_str] = day_types
+
     team.save()
     return {"team": mongo_to_pydantic(team, TeamReadDTO)}
 
 
+@app.put("/teams/{team_id}/members/{team_member_id}/days")
+def update_days(team_id: str, team_member_id: str, days: Dict[str, List[str]]):
+    """Assume it is an update for only one day"""
+    team: Team = Team.objects(id=team_id).first()
+    if not team:
+        raise HTTPException(status_code=404, detail="Team not found")
+
+    team_member: TeamMember = team.team_members.get(uid=team_member_id)
+    if not team_member:
+        raise HTTPException(status_code=404, detail="Team member not found")
+
+    # Convert the day type IDs to DayType references
+    updated_days = {}
+    for date_str, day_type_ids in days.items():
+        day_types = [DayType.objects(id=day_type_id).first() for day_type_id in day_type_ids]
+        updated_days[date_str] = sorted(day_types, key=lambda day_type: day_type.name)
+
+    team_member.days = team_member.days | updated_days
+    team.save()
+    return {"team": mongo_to_pydantic(team, TeamReadDTO)}
+
+
+@app.get("/daytypes/")
+def get_all_day_types():
+    day_types = DayType.objects.order_by("name")
+    return {"day_types": [mongo_to_pydantic(day_type, DayTypeReadDTO) for day_type in day_types]}
+
+
 @app.post("/daytypes/")
 def create_day_type(day_type_dto: DayTypeWriteDTO):
+    if not day_type_dto.color:
+        day_type_dto.color = None
     day_type_data = day_type_dto.model_dump()
     DayType(**day_type_data).save()
     return {"day_types": [mongo_to_pydantic(day_type, DayTypeReadDTO) for day_type in DayType.objects.order_by("name")]}
@@ -232,9 +280,25 @@ def update_day_type(day_type_id: str, day_type_dto: DayTypeWriteDTO):
     return {"day_types": [mongo_to_pydantic(day_type, DayTypeReadDTO) for day_type in DayType.objects.order_by("name")]}
 
 
+def flatten_list(list_of_lists):
+    return [item for sublist in list_of_lists for item in sublist]
+
+
 @app.delete("/daytypes/{day_type_id}")
 def delete_day_type(day_type_id: str):
+    # Check if DayType is used in any TeamMember's days or available_day_types, or in any Team's available_day_types
+    if any(
+            day_type_id in (str(day_types.id) for day_types in flatten_list(member.days.values()))
+            for team in Team.objects
+            for member in team.team_members
+    ):
+        raise HTTPException(
+            status_code=status.HTTP_400_BAD_REQUEST,
+            detail="DayType is in use and cannot be deleted")
+
+    # Proceed with deletion if DayType is not in use
     result = DayType.objects(id=day_type_id).delete()
     if result == 0:
         raise HTTPException(status_code=404, detail="DayType not found")
+
     return {"message": "DayType deleted successfully"}
